# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es una API para gestionar pacientes, dentistas y turnos en una clínica odontológica.",
        "title": "API de Gestión de Turnos en Clínica Odontológica",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/turnos": {
            "get": {
                "description": "Obtiene todos los turnos de un paciente específico por su DNI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Obtener turnos por DNI del paciente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DNI del Paciente",
                        "name": "dni",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Turno"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Agrega un nuevo turno al sistema, el paciente y dentista deben existir",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Crear un nuevo turno",
                "parameters": [
                    {
                        "description": "Objeto Turno",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    }
                }
            }
        },
        "/turnos/dm": {
            "post": {
                "description": "Agrega un nuevo turno al sistema usando DNI, Matrícula para búsqueda y también cargar fecha y descripción del turno",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Crear un nuevo turno por DNI y Matrícula",
                "parameters": [
                    {
                        "description": "Datos de la Solicitud",
                        "name": "requestData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RequestData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    }
                }
            }
        },
        "/turnos/{id}": {
            "get": {
                "description": "Obtiene los detalles de un turno específico por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Obtener un turno por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del Turno",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza un turno existente en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Actualizar un turno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del Turno",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Objeto Turno Actualizado",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un turno existente del sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Eliminar un turno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del Turno",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Actualiza parcialmente un turno existente en el sistema, solo se actualizan los campos proporcionados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "turnos"
                ],
                "summary": "Actualizar parcialmente un turno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del Turno",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Objeto Turno Parcial",
                        "name": "turno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Turno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/web.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Dentista": {
            "type": "object",
            "required": [
                "apellido",
                "matricula",
                "nombre"
            ],
            "properties": {
                "apellido": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "matricula": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "domain.Paciente": {
            "type": "object",
            "required": [
                "apellido",
                "direccion",
                "dni",
                "nombre"
            ],
            "properties": {
                "apellido": {
                    "type": "string"
                },
                "direccion": {
                    "type": "string"
                },
                "dni": {
                    "type": "string"
                },
                "fechaAlta": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "domain.RequestData": {
            "type": "object",
            "required": [
                "descripcion",
                "dni",
                "fecha",
                "matricula"
            ],
            "properties": {
                "descripcion": {
                    "type": "string"
                },
                "dni": {
                    "type": "string"
                },
                "fecha": {
                    "type": "string"
                },
                "matricula": {
                    "type": "string"
                }
            }
        },
        "domain.Turno": {
            "type": "object",
            "required": [
                "descripcion"
            ],
            "properties": {
                "dentista": {
                    "$ref": "#/definitions/domain.Dentista"
                },
                "descripcion": {
                    "type": "string"
                },
                "fecha": {
                    "description": "PacienteID int `json:\"pacienteId\" binding:\"required\"`\nDentistaID int `json:\"dentistaId\" binding:\"required\"`",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "paciente": {
                    "$ref": "#/definitions/domain.Paciente"
                }
            }
        },
        "web.errorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}